+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Implement support for GET /contacts/groups
 * Update lodash import
 * Support GET /resources
+* Support POST, PUT and DELETE for calendars, and add location, timezone and isPrimary attributes.
 
 ### 5.1.0 / 2020-06-04
 * Fix bug which was overwriting properties on message objects.